+++ hw3D/IndexBuffer.h
@@ -9,8 +9,8 @@ class IndexBuffer : public Bindable
 public:
 	IndexBuffer(Graphics& gfx, std::vector<unsigned short>& indices);
 	void Bind(Graphics& gfx) const override;
-	size_t GetCount() const;
+	uint32_t GetCount() const;
 private:
 	Microsoft::WRL::ComPtr<ID3D11Buffer> pData;
-	size_t count;
+	uint32_t count;
 };
